So forgot to actually test some situations.
backup_check - should now actually work with no arguments
backup - if sha256 can't be calculated, should report problem

+++ backup.sh
@@ -50,16 +50,16 @@ cp -rp $bkup_path/*.log $dest_path
 file1sha=(`shasum -a 256 $archive`)
 file2sha=(`shasum -a 256 $archivedest`)
 
-if [ "$file1sha" == "$file2sha" ]
+if [[ "$file1sha" == "$file2sha" && "$file1sha" != "" ]]
 then
     echo "Local and Remote files have the same content - removing local"
-    echo $archivedest $file1sha
+    echo "$archivedest has sha256: $file1sha"
     rm $archive
     rm $newmetadata
 else
-    echo "Files are NOT the same"
-    echo $archive $file1sha
-    echo $archivedest $file2sha
+    echo "Files are NOT the same or cannot calculate sha256"
+    echo "$archive has sha256: $file1sha"
+    echo "$archivedest has sha256: $file2sha"
 fi
 
 echo $(date)
